Gallery: Improve loading of media frame when editing a gallery.
Previously, we were replacing the media frame state right before opening
the modal, but after views like the menu were created. This caused a few
bugs, like the cancel button moving back to the initial "create gallery"
state and some labels being incorrect. This creates a selection from any
existing IDs and passes it to the frame before it is created to avoid
these issues.

Additionally, this updates the method for getting a collection of
attachments using `wp.media.query` so we can fetch all of the attachment
models in one request, rather than relying on seperate requests for each
attachment in the gallery.