Try adding a page heading
This is a work in progress, which adds a page heading for screen readers. @afercia can you take a look and see if this mitigates #503?

If yes, then we'd need to wrap the actual label into a function so it not only outputs the right post type, but indeed is translatable.